Refactor Maincontent component into app.js
